Require rails_helper for specs
