+++ spec/controllers/sites_controller_spec.rb
@@ -1,4 +1,4 @@
-require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
+require 'rails_helper'
 
 describe SitesController do
   before(:each) { login_as :jordan }
