+++ spec/controllers/users_controller_spec.rb
@@ -1,32 +1,32 @@
-require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
+require 'rails_helper'
 
 describe UsersController do
   before(:each) { login_as :jordan }
-  
+
   it 'handles /users with GET' do
     get :index
     response.should be_success
   end
-  
+
   it 'handles /users/:id with GET' do
     get :show, id: users(:jordan)
     response.should be_success
   end
-  
+
   it 'handles /users with valid params and POST' do
     running {
       post :create, user: {name: 'name', password: 'password'}
       response.should redirect_to(users_path)
     }.should change(User, :count).by(1)
   end
-  
+
   it 'handles /users/:id with valid params and PUT' do
     user = users(:jordan)
     put :update, id: user.to_param, user: {name: 'new'}
     user.reload.name.should == 'new'
     response.should redirect_to(user_path(user))
   end
-  
+
   it 'handles /users/:id with invalid params and PUT' do
     user = users(:jordan)
     put :update, id: user, user: {password: ''}
@@ -34,7 +34,7 @@ describe UsersController do
     response.should be_success
     response.should render_template(:show)
   end
-  
+
   it 'handles /users/:id with DELETE' do
     running {
       delete :destroy, id: users(:jordan)
