+++ spec/models/site_spec.rb
@@ -1,4 +1,4 @@
-require File.expand_path(File.dirname(__FILE__) + '/../spec_helper')
+require 'rails_helper'
 
 describe Site, 'relationships' do
   it 'should belong to a user' do
@@ -16,11 +16,11 @@ describe Site, 'validations' do
   it 'should have an url' do
     Site.new.should have(1).error_on(:url)
   end
-  
+
   it 'should have a valid url' do
     Site.new(url: 'invalid').should have(1).error_on(:url)
   end
-  
+
   it 'should belong to a user' do
     Site.new.should have(1).error_on(:user_id)
   end
